Separated LocalCollection.
