+++ packages/minimongo/cursor.js
@@ -1,0 +1,3 @@
+export class Cursor {
+
+}
