+++ packages/minimongo/observe_handle.js
@@ -1,0 +1,2 @@
+// the handle that comes back from observe.
+export class ObserveHandle {}
